galiano: removed most UI elements ready for viewshed ones
And a fair bit of refactoring
